% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_picarro_data.r
\name{choose_directory}
\alias{choose_directory}
\title{Interactive dialogue to allow user to select a directory}
\usage{
choose_directory(caption = "Select data directory")
}
\arguments{
\item{caption}{Optional text to print in title of selection dialogue}
}
\value{
Length one character vector containing path to selected directory
}
\description{
Function to select a directory interactively in a platform-independent way
}
\details{
This function allows directory selection in a way that should work 
on any (most?) platforms.
}
\examples{
choose_directory()
}
\author{
Dominic Woolf.
  d.woolf@cornell.edu
}
