% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_picarro_data.r
\name{extract_picarro}
\alias{extract_picarro}
\title{Data extraction from Lehmann lab Picarro isotope analyzer}
\usage{
extract_picarro(data_path = NA, lambda = 1e-04)
}
\arguments{
\item{data_path}{Directory containing all raw data from picarro analyzer within nested subdirectories. If not provided, user will be asked for path interactively.}

\item{lambda}{Optional smoothing parameter (positive real numeric). Higher values give greater smoothing. Smaller values follow data more closely. See help(regSmooth) for more details.}
}
\value{
Provides a data.table with respired CO2 and CH4 concentrations and delta-13C values for each jar at each samping time
}
\description{
Function to convert raw data from picarro analyzer to respiration data
}
\details{
This function returns a data.table with respired CO2 and CH4 
concentrations and delta-13C values for each jar at each samping time

The data path \strong{must} contain the following three things: 
\enumerate{
  \item All raw Picarro data files for this experiment. Can be in nested subdirectories.
  NB picarro files are assumed to end in .dat -- No other files in the data path should have this ending
  \item fractional_volume.txt  = provides fraction of head space gas divided by total volume including residal gas in the analyzer and pipework
  \item logfile (filename must start with the characters "logfile"), which contains epoch time, sample ID, and step ID 
          ("Step 2" = sample analysis;
           "Step 3" = end of step 2;
           "Step 5" = purge analysis;
           "Step 6" = end of step 5).
}
The returned data.table columns, and their descriptions are:

jar = jar number to which row corresponds
epoch = epoch time of measurement
cycle = number of sampling cycle (i.e. increments each time a specific jar is analyzed)
combined_jar_cycle = jar number and cycle number pasted together
CO2_purge = Picarro-measured CO2 concentration at the end of the purge step
CO2_respiration = Picarro-measured CO2 concentration at the peak (smoothed maximum) of the sampling step 
d13C_purge = Picarro-measured d13-CO2 at the end of the purge step
d13C_respiration = Picarro-measured d13-CO2 at the concentration-peak (smoothed maximum) of the sampling step 
<dilute.co2 is 1st value from each jar/cycle. I.e. concentration in pipework before headspace gas reaches the analyser>
dilute.co2_purge = dilute.co2 concentration in purge step
dilute.co2_respiration = dilute.co2 concentration in sampling step
dilute.d13C_purge = dilute.co2 d13C in purge step
dilute.d13C_respiration = dilute.co2 d13C in sampling step
fractional_volume = jar-specific correction factor for fraction of gas at picarro that originates from within the headspace rather than the pipework
headspace.co2 = CO2_respiration, corrected for fractional_volume
headspace.co2_purge = CO2_purge, corrected for fractional_volume
fractional_mass = fractional_volume converted to mass of CO2 basis
respired.co2 = CO2_respiration, corrected for fractional_volume
headspace.d13c = d13C_respiration, corrected for fractional_volume \strong{This correction provides highly suspect values.  Use d13C_respiration values instead}
day = epoch time converted into days since the first measurement on that sample
<Methane (CH4 has columns corresponding to the CO2 descriptions above):
   CH4_purge, CH4_respiration, d13CH4_purge, d13CH4_respiration, dilute.ch4_purge, 
   dilute.ch4_respiration, dilute.d13CH4_purge, dilute.d13CH4_respiration
   CH4_purge, headspace.ch4, headspace.ch4_purge, headspace.d13ch4, evolved.ch4
}
\examples{
short.data = extract_picarro()
library(ggplot2)
ggplot(short.data[-1], aes(day, respired.co2)) +
  geom_line() +
  facet_wrap(~ jar)

ggplot(short.data[-1], aes(day, evolved.ch4)) +
  geom_line() +
  facet_wrap(~ jar)
}
\author{
Dominic Woolf.
  d.woolf@cornell.edu
}
