% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_picarro_data.r
\name{partition_picarro}
\alias{partition_picarro}
\title{Isotope partitioning of CO2 and CH4 data from Picarro analyzer}
\usage{
partition_picarro(respiration_data = short.data, data_path = NA,
  remove_first = TRUE)
}
\arguments{
\item{respiration_data}{data.table containing extracted respiration data from Picarro analyzer. Typically created by function extract_picarro}

\item{data_path}{Directory containing jars.txt and treatments.txt.  If not provided, user will be asked for path interactively.}
}
\value{
Provides a data.table with the original respired CO2 and CH4 concentrations and delta-13C values 
for each jar at each samping time, from the input data.table. 
Columns are added to this data.table for mass of CO2 and CH4 respired during ech sampling period,
cumulative values of the same - both total and partitioned per substrate.
}
\description{
Function to partition respiration data from picarro analyzer by source, using d13C of two substrates.
Partitioning is done using the d13C values as measured by the picarro, without correction for fractional volume
}
\details{
This function returns a data.table with respired CO2 and CH4 
concentrations and delta-13C values for each jar at each samping time

The data path \strong{must} contain the following: 
\enumerate{
  \item jars.txt = file specifying the treatment in each jar. Also has columns for quantity of each substrate, and the headspace volume of the jar.
  \item d13c.txt = file specifying d13C of two substrates for each treatment. Treatment names must correspond with those in jars.txt
}
}
\examples{
short.data = extract_picarro()
partition.data = partition_picarro(short.data)
}
\author{
Dominic Woolf.
  d.woolf@cornell.edu
}
