% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_picarro_data.r
\name{extract_picarro}
\alias{extract_picarro}
\title{Data extraction from Lehmann lab Picarro isotope analyzer}
\usage{
extract_picarro(data_path = NA, lambda = 1e-04)
}
\arguments{
\item{data_path}{Directory containing all raw data from picarro analyzer within nested subdirectories. If not provided, user will be asked for path interactively.}

\item{lambda}{Optional smoothing parameter (positive real numeric). Higher values give greater smoothing. Smaller values follow data more closely. See help(regSmooth) for more details.}
}
\value{
Provides a data.table with respired CO2 and CH4 concentrations and delta-13C values for each sample at each samping time
}
\description{
Function to convert raw data from picarro analyzer to respiration data
}
\details{
This function returns a data.table with respired CO2 and CH4 concentrations and delta-13C values for each sample at each samping time

The data path **must** contain the following three things: 
  1. All raw Picarro data files nfor this experiment. Can be in nested subdirectories.
  NB picarro files are assumed to end in .dat -- No other files in the data path should have this ending
  2. fractional_volume.txt  = provides fraction of head space gas divided by total volume including residal gas in the analyzer and pipework
  3. logfile (filename must start with the characters "logfile"), which contains epoch time, sample ID, and step ID:
          Step 2 = sample analysis
          Step 3 = end of step 2
          Step 5 = purge analysis
          Step 6 = end of step 5
}
\examples{
short.data = extract_picarro()
library(ggplot2)
ggplot(short.data[-1], aes(day, respired.co2)) +
  geom_line() +
  facet_wrap(~ sample)

ggplot(short.data[-1], aes(day, evolved.ch4)) +
  geom_line() +
  facet_wrap(~ sample)
}
\author{
Dominic Woolf.
  d.woolf@cornell.edu
}
\keyword{Incubation}
\keyword{Soil}
